server/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
backports.tarfile==1.2.0
bidict==0.23.1
build==1.2.2.post1
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.5
djangorestframework==3.15.2
docutils==0.21.2
drf-nested-routers==0.94.1
fastapi==0.115.8
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
lazy_loader==0.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.20
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psutil==6.1.1
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
readme_renderer==44.0
redis==5.2.1
reflex==0.6.8
reflex-chakra==0.6.3
reflex-hosting-cli==0.1.35
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.45.3
starlette-admin==0.14.1
tabulate==0.9.0
tomlkit==0.13.2
twine==6.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
